Medical Triage System/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0
faiss-cpu>=1.7.4
reportlab>=4.0.0
requests>=2.31.0
